scripts/requirements.txt
yt-dlp
youtube-transcript-api
google-generativeai
supabase
python-dotenv
